+++ TextExpansion.cs
@@ -5,6 +5,9 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
 using System.Windows.Threading;
+using DataGrid = System.Windows.Controls.DataGrid;
+using FlowDirection = System.Windows.FlowDirection;
+using TextBox = System.Windows.Controls.TextBox;
 
 namespace WinVClipboard;